requirements.txt
elasticsearch>=6.2.0
tweepy
python-dotenv
